Fix schema dump when table contains both a regular and context index
When dumping a table schema that contains both a regular index and a context index,
it would dump an error to `schema.rb`.

```
Could not dump table "posts" because of following ArgumentError
  comparison of NilClass with String failed
```

The problem is when you had both types of indexes, the array would include a `nil`
which would blow up the sort.

Fixes #1985